+++ packages/cli/CHANGELOG.md
@@ -2,6 +2,14 @@
 
 ## Unreleased
 
+### Changed
+
+- Commands `add`, `remove` and `update` run codegen by default after finishing [#577](https://github.com/polkadot-api/polkadot-api/pull/577).
+
+### Removed
+
+- Remove `key` argument from `generate` [#577](https://github.com/polkadot-api/polkadot-api/pull/577).
+
 ### Fixed
 
 - Whitelist entry for pallet wildcard